refactor: clean & split xcm tests
